chore: change overdue days from long to int

+++ src/main/java/com/DevSprint/LibraryMS/dto/LendingDTO.java
@@ -16,6 +16,6 @@ public class LendingDTO implements Serializable{
     private String lendingDate;
     private String returnDate;
     private Boolean isActiveLending;
-    private Long overdueDays;
+    private Integer overdueDays;
     private Double fineAmount;
 }
